Making things more data-driven and pulling out reused code
