+++ client/src/components/ManageQuestions.vue
@@ -11,50 +11,29 @@
         <new-question-form @add="addNewQuestion"></new-question-form>
         <hr>
         <h2> Edit Current Questions </h2>
-        <question v-for="question in latest_question_list" :question="question" :key="question.id" :showEditControls="true"></question>
+        <question v-for="question in latest_question_list" :question="question" :key="question.id" :showEditControls="true" @delete="deleteQuestion"></question>
     </div>
 </template>
 
 <script>
-import axios from 'axios';
 import NewQuestionForm from './NewQuestionForm.vue';
 import Question from './Question.vue';
+import QuestionMixin from './../mixins/questions.js';
 export default {
-    data(){
-        return {
-            latest_question_list: []
-        }
-    },
     components: {NewQuestionForm, Question},
-    created: function(){
-        axios.get('http://localhost:8000/polls/questions')
-        .then((response) => {
-            this.latest_question_list = response.data.map((question)=> {
-                return {
-                    id: question.id,
-                    expanded: false,
-                    editing: false,
-                    visible: true,
-                    question_text: question.question_text
-                }
-            });
-        })
-        .catch((err) => console.log(err));
-    },
-    destroyed() {
-        this.$destroy();
-    },
+    mixins: [QuestionMixin],
     methods: {
         addNewQuestion(response){
             const data = response.data;
-            this.latest_question_list.unshift({
+            this.latest_question_list = [{
                 id: data.id,
                 expanded: false,
                 editing: false,
-                visible: true,
                 question_text: data.question_text
-            });
-
+            }, ...this.latest_question_list];
+        },
+        deleteQuestion(questionId){
+            this.latest_question_list = this.latest_question_list.filter((question)=> question.id != questionId);
         }
     }
 }
