+++ client/src/components/Polls.vue
@@ -14,37 +14,11 @@
 </template>
 
 <script>
-    import axios from 'axios';
     import Question from './Question.vue';
+    import QuestionMixin from '../mixins/questions.js';
     export default {
-        data(){
-            return {
-                latest_question_list: []
-            }
-        },
         components: {Question},
-        created: function(){
-            axios.get('http://localhost:8000/polls/questions')
-            .then((response) => {
-                this.latest_question_list = response.data.map((question)=> {
-                    return {
-                        id: question.id,
-                        expanded: false,
-                        visible: true,
-                        question_text: question.question_text
-                    }
-                });
-            })
-            .catch((err) => console.log(err));
-        },
-        destroyed() {
-            this.$destroy();
-        },
-        computed: {
-            detail_url: function (id) {
-                return 'http://localhost:8000/questions/' + id
-            }
-        }
+        mixins: [QuestionMixin]
     }
 </script>
